[Delete Confirm]
- add delete confirm dialog call function
  - ecrf-user.js

- create delete exp group action command

- add delete confirm dialog language key
  - title language key
  - content language key

- add delete confirm dialog
  - at list delete button
  - at update delete button

- comment some log
  - ListSubjectRenderCommand.java
  - ResearcherPermission.java

- add menu check code
  - researcher update menu

- change button shadow attribute

+++ ecrf-user-crf/src/main/resources/META-INF/resources/html/crf-data/dialog/crf-data-selector.jsp
@@ -233,6 +233,26 @@ function toEachCRF(sdId, isAudit){
 }
 
 function deleteEachCRF(linkCrfId){
-	Liferay.Util.getOpener().deleteCRFData(linkCrfId, <%=crfId%>, '<portlet:namespace/>multiCRFDialog', '<%=themeDisplay.getPortletDisplay().getId() %>', <%=themeDisplay.getPlid() %>, '<%=baseURL%>');	
+	// cant use general confirm dialog code, need to change ok action function
+	$.confirm({
+		title: '<liferay-ui:message key="ecrf-user.message.confirm-delete-crf-data.title"/>',
+		content: '<p><liferay-ui:message key="ecrf-user.message.confirm-delete-crf-data.content"/></p>',
+		type: 'red',
+		typeAnimated: true,
+		columnClass: 'large',
+		buttons:{
+			ok: {
+				btnClass: 'btn-blue',
+				action: function(){
+					Liferay.Util.getOpener().deleteCRFData(linkCrfId, <%=crfId%>, '<portlet:namespace/>multiCRFDialog', '<%=themeDisplay.getPortletDisplay().getId() %>', <%=themeDisplay.getPlid() %>, '<%=baseURL%>');				
+				}
+			},
+			close: {
+	            action: function () {
+	            }
+	        }
+		},
+		draggable: true
+	});
 }
 </script>
